+++ src/components/property-page-widgets/RoomType/utils/getModalContentMarkup.js
@@ -5,7 +5,6 @@ import { buildKeyFromStrings } from 'utils/build-key-from-strings';
 import { Amenities } from 'property-page-widgets/Amenities';
 import { Divider } from 'elements/Divider';
 import { Paragraph } from 'typography/Paragraph';
-import { Button } from 'elements/Button';
 import { Grid } from 'layout/Grid';
 import { GridColumn } from 'layout/GridColumn';
 import { Heading } from 'typography/Heading';
@@ -14,7 +13,6 @@ import { Slideshow } from 'media/Slideshow';
 
 /**
  * @param  {Object[]}    amenities
- * @param  {Function}    onClickCheckAvailability
  * @param  {string}      description
  * @param  {Object[]}    extraFeatures
  * @param  {Object[]}    features
@@ -22,18 +20,19 @@ import { Slideshow } from 'media/Slideshow';
  * @param  {string}      nightPrice
  * @param  {number}      ratingNumber
  * @param  {Object[]}    slideShowImages
+ * @param  {boolean}     isUserOnMobile
  * @return {Object}
  */
 export const getModalContentMarkup = (
   amenities,
-  onClickCheckAvailability,
   description,
   extraFeatures,
   features,
   name,
   nightPrice,
   ratingNumber,
-  slideShowImages
+  slideShowImages,
+  isUserOnMobile
 ) => (
   <Modal.Content>
     <Heading>{name}</Heading>
@@ -49,19 +48,18 @@ export const getModalContentMarkup = (
       ))}
     </List>
     <Divider hasLine />
-    <Amenities amenities={amenities} headingText="Room Amenities" />
+    <Amenities
+      amenities={amenities}
+      headingText="Room Amenities"
+      isStacked={isUserOnMobile}
+    />
     <Grid>
-      <GridColumn verticalAlignContent="bottom" width={6}>
+      <GridColumn verticalAlignContent="bottom" width={12}>
         <Paragraph>
           from
           <strong>{` ${nightPrice} `}</strong>
           /night
         </Paragraph>
-      </GridColumn>
-      <GridColumn verticalAlignContent="bottom" width={6}>
-        <Button isPositionedRight isRounded onClick={onClickCheckAvailability}>
-          Check Availability
-        </Button>
       </GridColumn>
     </Grid>
   </Modal.Content>